test(e2e): isolate credential probe fixture
