+++ test/e2e/live/channels-stop-start-config-state.ts
@@ -37,20 +37,6 @@ export function openClawChannelIsInert(state: OpenClawChannelConfigState): boole
   return !state.channelEnabled && !state.channelHasEnabledAccount && channelIsInert;
 }
 
-const HERMES_REVISION_SCOPED_CREDENTIALS = {
-  wechat: ["WEIXIN_TOKEN", "WECHAT_BOT_TOKEN"],
-  teams: ["TEAMS_CLIENT_SECRET", "MSTEAMS_APP_PASSWORD"],
-} as const;
-
-export type HermesRevisionScopedCredentialChannel = keyof typeof HERMES_REVISION_SCOPED_CREDENTIALS;
-
-export function hermesRevisionScopedCredentialLinePattern(
-  channel: HermesRevisionScopedCredentialChannel,
-): string {
-  const [targetEnvKey, credentialEnvKey] = HERMES_REVISION_SCOPED_CREDENTIALS[channel];
-  return `^${targetEnvKey}=openshell:resolve:env:v[0-9]+_${credentialEnvKey}$`;
-}
-
 export function openClawChannelStateProbeScript(
   channel: string,
   configPath = "/sandbox/.openclaw/openclaw.json",
