+++ test/e2e/mock-parity.json
@@ -633,12 +633,6 @@
     },
     {
       "live": "test/e2e/live/channels-stop-start.test.ts",
-      "liveSources": [
-        "test/e2e/live/channels-stop-start-config-state.ts",
-        "test/e2e/live/channels-stop-start-googlechat-proof.ts",
-        "test/e2e/live/channels-stop-start-helpers.ts",
-        "test/e2e/live/channels-stop-start-plan-state.ts"
-      ],
       "fast": [
         "test/e2e/support/channels-stop-start-config-state.test.ts",
         "test/e2e/support/channels-stop-start-googlechat-entry.test.ts",
